move ServerOptions from util to main
